+++ runway/module/serverless.py
@@ -1,6 +1,7 @@
 """Serverless module."""
 from __future__ import print_function
 
+import argparse
 import logging
 import os
 import re
@@ -8,15 +9,17 @@ import shutil
 import subprocess
 import sys
 import tempfile
+import uuid
+
 import yaml
 
 from runway.hooks.staticsite.util import get_hash_of_files
-from . import (
-    RunwayModule, format_npm_command_for_logging, generate_node_command,
-    run_module_command, run_npm_install, warn_on_boto_env_vars
-)
-from ..util import change_dir, which
-from ..s3_util import ensure_bucket_exists, does_s3_object_exist, download, upload
+
+from ..s3_util import (does_s3_object_exist, download, ensure_bucket_exists,
+                       upload)
+from ..util import cached_property, merge_dicts
+from . import (ModuleOptions, RunwayModuleNpm, format_npm_command_for_logging,
+               generate_node_command, run_module_command)
 
 LOGGER = logging.getLogger('runway')
 
@@ -40,32 +43,9 @@ def gen_sls_config_files(stage, region):
     return names
 
 
-def get_sls_config_file(path, stage, region):
-    """Determine Serverless config file name."""
-    for name in gen_sls_config_files(stage, region):
-        if os.path.isfile(os.path.join(path, name)):
-            return name
-    return "config-%s.json" % stage  # fallback to generic json name
-
-
-def run_sls_remove(sls_cmd, env_vars):
-    """Run sls remove command."""
-    sls_process = subprocess.Popen(sls_cmd,
-                                   stdout=subprocess.PIPE,
-                                   env=env_vars)
-    stdoutdata, _stderrdata = sls_process.communicate()
-    sls_return = sls_process.wait()
-    if int(sys.version[0]) > 2:
-        stdoutdata = stdoutdata.decode('UTF-8')  # bytes -> string
-    print(stdoutdata)
-    if sls_return != 0 and (sls_return == 1 and not (
-            re.search(r"Stack '.*' does not exist", stdoutdata))):
-        sys.exit(sls_return)
-
-
 def run_sls_print(sls_opts, env_vars, path):
     """Run sls print command."""
-    sls_info_opts = sls_opts
+    sls_info_opts = list(sls_opts)
     sls_info_opts[0] = 'print'
     sls_info_opts.extend(['--format', 'yaml'])
     sls_info_cmd = generate_node_command(command='sls',
@@ -93,12 +73,9 @@ def get_src_hash(sls_config, path):
 
     return hashes
 
-def deploy_package(sls_opts, options, context, path): # noqa pylint: disable=too-many-locals
-    """Run sls package command."""
-    bucketname = options.get('options', {}).get('promotezip', {}).get('bucketname', {})
-    if not bucketname:
-        raise ValueError('"bucketname" must be specified when using "promotezip"')
 
+def deploy_package(sls_opts, bucketname, context, path):
+    """Run sls package command."""
     package_dir = tempfile.mkdtemp()
     LOGGER.debug('Package directory: %s', package_dir)
 
@@ -107,8 +84,7 @@ def deploy_package(sls_opts, options, context, path): # noqa pylint: disable=too
     hashes = get_src_hash(sls_config, path)
 
     sls_opts[0] = 'package'
-    sls_opts.extend(['--package', os.path.relpath(package_dir,
-                                                  path)])
+    sls_opts.extend(['--package', package_dir])
     sls_package_cmd = generate_node_command(command='sls',
                                             command_opts=sls_opts,
                                             path=path)
@@ -146,89 +122,344 @@ def deploy_package(sls_opts, options, context, path): # noqa pylint: disable=too
     shutil.rmtree(package_dir)
 
 
-class Serverless(RunwayModule):
+class Serverless(RunwayModuleNpm):
     """Serverless Runway Module."""
 
-    def run_serverless(self, command='deploy'):
-        """Run Serverless."""
-        response = {'skipped_configs': False}
-        sls_opts = [command]
+    def __init__(self, context, path, options=None):
+        """Instantiate class.
 
-        if not which('npm'):
-            LOGGER.error('"npm" not found in path or is not executable; '
-                         'please ensure it is installed correctly.')
+        Args:
+            context (Context): Runway context object.
+            path (str): Path to the module.
+            options (Dict[str, Dict[str, Any]]): Everything in the module
+                definition merged with applicable values from the deployment
+                definition.
+
+        """
+        options = options or {}
+        super(Serverless, self).__init__(context, path, options.copy())
+        try:
+            self.options = ServerlessOptions.parse(**options.get('options',
+                                                                 {}))
+        except ValueError as err:
+            LOGGER.error('%s: %s', self.path.name, err)
             sys.exit(1)
+        self.region = self.context.env_region
+        self.stage = self.context.env_name
 
-        if 'CI' in self.context.env_vars and command != 'remove':
-            sls_opts.append('--conceal')  # Hide secrets from serverless output
+    @property
+    def cli_args(self):
+        """Generate CLI args from self used in all Serverless commands.
 
+        Returns:
+            List[str]
+
+        """
+        result = ['--region', self.region,
+                  '--stage', self.stage]
         if 'DEBUG' in self.context.env_vars:
-            sls_opts.append('-v')  # Increase logging if requested
+            result.append('--verbose')
+        return result
 
-        warn_on_boto_env_vars(self.context.env_vars)
+    @cached_property
+    def env_file(self):
+        """Find the environment file for the module.
 
-        sls_opts.extend(['-r', self.context.env_region])
-        sls_opts.extend(['--stage', self.context.env_name])
-        sls_env_file = get_sls_config_file(self.path,
-                                           self.context.env_name,
-                                           self.context.env_region)
-        sls_opts.extend(self.options.get('options', {}).get('args', []))
+        Returns:
+            Path: Path object for the environment file that was found.
 
-        sls_cmd = generate_node_command(command='sls',
-                                        command_opts=sls_opts,
-                                        path=self.path)
+        """
+        for name in gen_sls_config_files(self.stage, self.region):
+            test_path = self.path / name
+            if test_path.is_file():
+                return test_path
+        return None
 
-        if (
-                self.options['parameters'] or
-                os.path.isfile(os.path.join(self.path, sls_env_file))
-        ):
-            if os.path.isfile(os.path.join(self.path, 'package.json')):
-                with change_dir(self.path):
-                    run_npm_install(self.path, self.options, self.context)
-                    if command == 'deploy' and self.options.get('options', {}).get('promotezip', {}): # noqa pylint: disable=line-too-long
-                        deploy_package(sls_opts,
-                                       self.options,
-                                       self.context,
-                                       self.path)
-                        return response
+    @property
+    def skip(self):
+        """Determine if the module should be skipped.
 
-                    LOGGER.info("Running sls %s on %s (\"%s\")",
-                                command,
-                                os.path.basename(self.path),
-                                format_npm_command_for_logging(sls_cmd))
-                    if command == 'remove':
-                        # Need to account for exit code 1 on any removals after
-                        # the first
-                        run_sls_remove(sls_cmd, self.context.env_vars)
-                    else:
-                        run_module_command(cmd_list=sls_cmd,
-                                           env_vars=self.context.env_vars)
-            else:
-                LOGGER.warning(
-                    "Skipping serverless %s of %s; no \"package.json\" "
-                    "file was found (need a package file specifying "
-                    "serverless in devDependencies)",
-                    command,
-                    os.path.basename(self.path))
+        Returns:
+            bool: To skip, or not to skip, that is the question.
+
+        """
+        if not self.package_json_missing():
+            if self.parameters or self.environments or self.env_file:
+                return False
+            LOGGER.warning('%s: No config file for this stage/region found '
+                           '(looking for one of "%s")', self.path.name,
+                           ', '.join(gen_sls_config_files(self.stage,
+                                                          self.region)))
         else:
-            response['skipped_configs'] = True
-            LOGGER.info(
-                "Skipping serverless %s of %s; no config file for "
-                "this stage/region found (looking for one of \"%s\")",
-                command,
-                os.path.basename(self.path),
-                ', '.join(gen_sls_config_files(self.context.env_name,
-                                               self.context.env_region)))
-        return response
+            LOGGER.warning('%s: The Serverless module type requires a package '
+                           'file specifying serverless in devDependencies',
+                           self.path.name)
+        LOGGER.warning('%s: Skipping module', self.path.name)
+        return True
+
+    def extend_serverless_yml(self, func):
+        """Extend the Serverless config file with additional YAML from options.
+
+        Args:
+            func (Callable): Callable to use after handling the Serverless
+                config file.
+
+        """
+        self.npm_install()  # doing this here for a cleaner log
+        LOGGER.info('%s: Extending Serverless config from runway.yml...',
+                    self.path.name)
+        final_yml = merge_dicts(self.sls_print(skip_install=True),
+                                self.options.extend_serverless_yml)
+        # using a unique name to prevent collisions when run in parallel
+        tmp_file = self.path / '{}.tmp.serverless.yml'.format(uuid.uuid4())
+
+        try:
+            LOGGER.debug('%s: Creating temporary serverless config... "%s"',
+                         self.path.name, tmp_file.name)
+            if self.context.is_python3:
+                tmp_file.write_text(yaml.safe_dump(final_yml))
+            else:  # TODO remove handling when dropping python 2 support
+                tmp_file.write_text(yaml.safe_dump(final_yml).decode('UTF-8'))
+            # update args from options with the new config name
+            self.options.update_args('config', str(tmp_file.name))
+            func(skip_install=True)
+        finally:
+            try:
+                tmp_file.unlink()  # always cleanup the temp file
+            except OSError as err:
+                # catch error raised if file does not exist but this is fine
+                LOGGER.debug('%s: %s', self.path.name, err)
+
+    def gen_cmd(self, command, args_list=None):
+        """Generate and log a Serverless command.
+
+        This does not execute the command, only prepares it for use.
+
+        Args:
+            command (str): The Serverless command to be executed.
+            args_list (Optiona[List[str]]): Additional arguments to include
+                in the generated command.
+
+        Returns:
+            List[str]: The full command to be passed into a subprocess.
+
+        """
+        args = [command] + self.cli_args + self.options.args
+        args.extend(args_list or [])
+        if command not in ['remove', 'print'] and self.context.is_noninteractive:
+            args.append('--conceal')  # hide secrets from serverless output
+        cmd = generate_node_command(command='sls',
+                                    command_opts=args,
+                                    path=self.path)
+        self.log_npm_command(cmd)
+        return cmd
+
+    def sls_deploy(self, skip_install=False):
+        """Execute ``sls deploy`` command.
+
+        Args:
+            skip_install (bool): Skip ``npm install`` before running the
+                Serverless command. (*default:* ``False``)
+
+        """
+        if not skip_install:
+            self.npm_install()
+
+        if self.options.promotezip:
+            # TODO refactor deploy_package to be part of the class
+            self.path.absolute()
+            deploy_package(['deploy'] + self.cli_args + self.options.args,
+                           self.options.promotezip['bucketname'],
+                           self.context,
+                           str(self.path))
+            return
+        run_module_command(cmd_list=self.gen_cmd('deploy'),
+                           env_vars=self.context.env_vars)
+
+    def sls_print(self, item_path=None, skip_install=False):
+        """Execute ``sls print`` command.
+
+        Keyword Args:
+            item_path (Optional[str]): Period-separated path to print a
+                sub-value (eg: "provider.name").
+            skip_install (bool): Skip ``npm install`` before running the
+                Serverless command. (*default:* ``False``)
+
+        Returns:
+            Dict[str, Any]: Resolved Serverless config file.
+
+        Raises:
+            SystemExit: If a runway-tmp.serverless.yml file already exists.
+
+        """
+        if not skip_install:
+            self.npm_install()
+
+        args = ['--format', 'yaml']
+        if item_path:
+            args.extend(['--path', item_path])
+        return yaml.safe_load(subprocess.check_output(
+            self.gen_cmd('print', args_list=args),
+            env=self.context.env_vars
+        ))
+
+    def sls_remove(self, skip_install=False):
+        """Execute ``sls remove`` command.
+
+        Args:
+            skip_install (bool): Skip ``npm install`` before running the
+                Serverless command. (*default:* ``False``)
+
+        """
+        if not skip_install:
+            self.npm_install()
+        stack_missing = False  # track output for acceptable error
+        proc = subprocess.Popen(self.gen_cmd('remove'),
+                                bufsize=1,
+                                env=self.context.env_vars,
+                                stdout=subprocess.PIPE,
+                                universal_newlines=True)
+        with proc.stdout:  # live output
+            for line in proc.stdout:
+                print(line, end='')
+                if re.search(r"Stack '.*' does not exist", line):
+                    stack_missing = True
+        if proc.wait() != 0 and not stack_missing:
+            sys.exit(proc.returncode)
 
     def plan(self):
-        """Skip sls planning."""
+        """Entrypoint for Runway's plan action."""
         LOGGER.info('Planning not currently supported for Serverless')
 
     def deploy(self):
-        """Run sls deploy."""
-        self.run_serverless(command='deploy')
+        """Entrypoint for Runway's deploy action."""
+        if self.skip:
+            return
+        if self.options.extend_serverless_yml:
+            self.extend_serverless_yml(self.sls_deploy)
+        else:
+            self.sls_deploy()
 
     def destroy(self):
-        """Run serverless remove."""
-        self.run_serverless(command='remove')
+        """Entrypoint for Runway's destroy action."""
+        if self.skip:
+            return
+        if self.options.extend_serverless_yml:
+            self.extend_serverless_yml(self.sls_remove)
+        else:
+            self.sls_remove()
+
+
+class ServerlessOptions(ModuleOptions):
+    """Module options for Serverless."""
+
+    def __init__(self, args, extend_serverless_yml, promotezip,
+                 skip_npm_ci=False):
+        """Instantiate class.
+
+        Keyword Args:
+            args (List[str]): Arguments to append to Serverless CLI
+                commands. These will always be placed after the default
+                arguments provided by Runway.
+            extend_serverless_yml (Dict[str, Any]): If provided,
+                a temporary Serverless config will be created will be created
+                from what exists in the module directory then the value of
+                this option will be merged into it. The temporary file will
+                be deleted at the end of execution.
+            promotezip (Dict[str, str]): If provided, promote
+                Serverless generated zip files between environments from a
+                *build* AWS account.
+            skip_npm_ci (bool): Skip the ``npm ci`` Runway executes at the
+                begining of each Serverless module run.
+
+        """
+        super(ServerlessOptions, self).__init__()
+        self._arg_parser = self._create_arg_parser()
+        self.extend_serverless_yml = extend_serverless_yml
+        cli_args, self._unknown_cli_args = \
+            self._arg_parser.parse_known_args(list(args)  # python 2 compatible
+                                              if isinstance(args, list)
+                                              else [])
+        self._cli_args = vars(cli_args)  # convert argparse.Namespace to dict
+        self.promotezip = promotezip
+        self.skip_npm_ci = skip_npm_ci
+
+    @property
+    def args(self):
+        """Args to pass to the CLI.
+
+        Returns:
+            List[str]: List of arguments.
+
+        """
+        known_args = []
+        for key, val in self._cli_args.items():
+            if isinstance(val, str):
+                known_args.extend(['--%s' % key, val])
+        return known_args + self._unknown_cli_args
+
+    def update_args(self, key, value):
+        """Update a known CLI argument.
+
+        Args:
+            key (str): Dict key to be updated.
+            value (str): New value
+
+        Raises:
+            KeyError: The key provided for update is now a known arg.
+
+        """
+        if key in self._cli_args:
+            self._cli_args[key] = value
+        else:
+            raise KeyError(key)
+
+    @staticmethod
+    def _create_arg_parser():
+        """Create argparse parser to parse args.
+
+        Used to pull arguments out of self.args when logic could change
+        depending on values provided.
+
+        Returns:
+            argparse.ArgumentParser
+
+        """
+        parser = argparse.ArgumentParser()
+        parser.add_argument('-c', '--config', default=None)
+        return parser
+
+    @classmethod
+    def parse(cls, **kwargs):  # pylint: disable=arguments-differ
+        """Parse the options definition and return an options object.
+
+        Keyword Args:
+            args (Optional[List[str]]): Arguments to append to Serverless CLI
+                commands. These will always be placed after the default
+                arguments provided by Runway.
+            extend_serverless_yml (Optional[Dict[str, Any]]): If provided,
+                a temporary Serverless config will be created will be created
+                from what exists in the module directory then the value of
+                this option will be merged into it. The temporary file will
+                be deleted at the end of execution.
+            promotezip (Optional[Dict[str, str]]): If provided, promote
+                Serverless generated zip files between environments from a
+                *build* AWS account.
+            skip_npm_ci (bool): Skip the ``npm ci`` Runway executes at the
+                begining of each Serverless module run.
+
+        Returns:
+            ServerlessOptions
+
+        Raises:
+            ValueError: promotezip was provided but missing bucketname.
+
+        """
+        promotezip = kwargs.get('promotezip', {})
+        if promotezip and not promotezip.get('bucketname'):
+            raise ValueError('"bucketname" must be provided when using '
+                             '"options.promotezip": {}'.format(promotezip))
+        return cls(args=kwargs.get('args', []),
+                   extend_serverless_yml=kwargs.get('extend_serverless_yml', {}),
+                   promotezip=promotezip,
+                   skip_npm_ci=kwargs.get('skip_npm_ci', False))